+++ pyincore/analyses/cumulativebuildingdamage/cumulativebuildingdamage.py
@@ -14,7 +14,11 @@ from pyincore import BaseAnalysis, AnalysisUtil
 
 
 class CumulativeBuildingDamage(BaseAnalysis):
-    """Cumulative building damage analysis."""
+    """This analysis computes the cumulative building damage for a combined earthquake and tsunami event.
+    The process for computing the structural damage is done externally and the results for earthquake and
+    tsunami are passed to this analysis. The damage intervals are then calculated from combined limit
+    state probabilities for the two hazards.
+    """
 
     def run(self):
         """Executes Cumulative Building Damage Analysis"""