+++ CHANGELOG.md
@@ -10,6 +10,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added 
 - Building direct, structural economic loss analysis.  [INCORE1-442](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-442).
 
+### Removed
+- Clean up redundant method in analysisutil [INCORE1-732](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-732)
+
+### Added
+- \_\_version__ property to show pyincore version. IncoreClient will print the detected version [INCORE1-520](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-520)
+- Building Functionality analysis outputs functionality samples as an additional result [INCORE1-734](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-734)
+- Python method to interact with flood endpoints in hazard service [INCORE1-747](https://opensource.ncsa.illinois
+.edu/jira/browse/INCORE1-747)
+### Changed
+- Calculate multiple limit states of custom expression fragility curves [INCORE1-682](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-682) 
+
+
+### Changed
+- Updated docstrings to include all hazards that each support [INCORE1-708](https://opensource.ncsa.illinois.edu/jira/browse/INCORE1-708)
+>>>>>>> develop
+
 ## [0.7.0] - 2020-07-31
 
 ### Added